requirements.txt
asgiref==3.8.1
attrs==23.1.0
boto3==1.26.138
botocore==1.29.138
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.2
defusedxml==0.7.1
dj-database-url==2.0.0
Django==5.2
django-cors-headers==4.0.0
django-debug-toolbar==4.1.0
django-environ==0.11.2
django-filter==23.5
django-rest-framework==0.1.0
django-storages==1.13.2
djangorestframework==3.14.0
djangorestframework_simplejwt==5.5.0
djoser==2.3.1
drf-spectacular==0.26.5
drf-spectacular-sidecar==2023.12.1
gunicorn==20.1.0
idna==3.10
inflection==0.5.1
jmespath==1.0.1
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
Markdown==3.7
oauthlib==3.2.2
Pillow==9.5.0
psycopg2==2.9.9
pycparser==2.22
PyJWT==2.9.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0.1
referencing==0.31.1
requests==2.32.3
requests-oauthlib==2.0.0
rpds-py==0.13.2
s3transfer==0.6.1
setuptools==69.2.0
six==1.16.0
social-auth-app-django==5.4.3
social-auth-core==4.5.6
sqlparse==0.4.4
typing_extensions==4.6.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==1.26.16
whitenoise==6.4.0
